+++ src/ui/about.rs
@@ -70,8 +70,9 @@ impl SimpleComponent for AboutDialog {
                 "Français — @zeGolem https://github.com/zeGolem",
                 "Türkçe — @Kaozix https://github.com/Kaozix1776",
                 "Türkçe — Kayra Nachfolger https://github.com/kayranachfolger",
-                "简体中文 — Caibin Chen https://github.com/tigersoldier",
-                "Italiano - @QuazarOmega https://github.com/quazar-omega"
+                "Italiano - @QuazarOmega https://github.com/quazar-omega",
+                "Indonesia - @yumekarisu https://github.com/yumekarisu",
+                "简体中文 — Caibin Chen https://github.com/tigersoldier"
             ].join("\n"),
 
             set_debug_info: &[
@@ -89,25 +90,43 @@ impl SimpleComponent for AboutDialog {
                 "<p>Added</p>",
 
                 "<ul>",
-                    "<li>Added rules approving dialog to the first run window</li>",
-                    "<li>Added game settings section</li>",
-                    "<li>Added game sessions manager</li>",
-                    "<li>Added `LAUNCHER_FOLDER` variable support</li>",
-                    "<li>Added patch repository mirror</li>",
-                "</ul>",
-
-                "<p>Changed</p>",
-
-                "<ul>",
-                    "<li>Improved launcher logo rendering quality</li>",
-                    "<li>Reworked entry rows in the settings</li>",
+                    "<li>Added Italian</li>",
+                    "<li>Added Indonesian</li>",
+                    "<li>Added dynamic main button icon switching</li>",
+                    "<li>Set button label as \"Resume\" when the diff is part downloaded</li>",
+                    "<li>Added options to use wine / gstreamer shared libraries from selected wine build. These options will configure `LD_LIBRARY_PATH` and `GST_PLUGIN_PATH` environment variables</li>",
+                    "<li>Added setting of `LC_ALL` in wine lang setting</li>",
+                    "<li>Added `LAUNCHER_REQUESTS_TIMEOUT` environment variable</li>",
                 "</ul>",
 
                 "<p>Fixed</p>",
 
                 "<ul>",
-                    "<li>Fixed wine tools running using proton builds</li>",
-                    "<li>Fixed sandboxed game running (sounds are broken for now)</li>",
+                    "<li>Fixed session applying on each launcher start</li>",
+                    "<li>Fixed predownload button ui</li>",
+                    "<li>Fixed proton builds integration with sandbox</li>",
+                    "<li>Fixed compatibility between sessions manager and sandbox</li>",
+                    "<li>Fixed sandboxing of inexisting folders</li>",
+                "</ul>",
+
+                "<p>Changed</p>",
+
+                "<ul>",
+                    "<li>Apply selected session before launching the game. This will properly save your game session when you switch between wine prefixes</li>",
+                    "<li>Redesigned main button</li>",
+                    "<li>Used `whatadistro` to identify recommended package manager in the first run window</li>",
+                    "<li>Moved a lot of settings to separate page</li>",
+                    "<li>Set fsr quality mode in enhancements settings instead of strength</li>",
+                    "<li>Updated fps unlocker data</li>",
+                    "<li>Made temporary workaround to the game API changes</li>",
+                    "<li>Increased default requests timeout to 8 seconds</li>",
+                    "<li>Updated minreq to support `http_proxy`-like variables</li>",
+                "</ul>",
+
+                "<p>Removed</p>",
+
+                "<ul>",
+                    "<li>Removed Futex2 wine sync option</li>",
                 "</ul>",
             ].join("\n"),